chatbot/requirements.txt
Flask==2.3.3
twilio==8.10.0
python-dotenv==1.0.1
openai==1.30.1
requests==2.31.0
